reorganize to reduce rerenders
