+++ src/Components/Filters/AddFilters/RenderCheckContainer.js
@@ -2,14 +2,13 @@ import React from "react";
 import CategoryOptions from "./CategoryOptions";
 import styled from "styled-components";
 import RadioGroup from "@material-ui/core/RadioGroup";
-//component - diplays a button to see options when button is clicked
+
 const RenderCheckContainer = ({
   i,
   filters,
   graphLabels,
   itemName,
   setFilters,
-  FilterBoxOptions,
   setUpdateUrlFlag,
   updateUrlFlag,
   setDisplayDrop
